[REEF-866] Refactor bridge clients to separate driver config generation from application submission
JIRA:
  [REEF-866](https://issues.apache.org/jira/browse/REEF-866)
